feat: refactor plugin types (#285)
