+++ packages/plugin-core/src/types/base.ts
@@ -2,4 +2,5 @@ export interface IExportData {
   specifier?: string;
   source: string;
   exportName: string;
+  extraExport?: boolean;
 }
