+++ packages/plugin-core/src/utils/checkExportData.ts
@@ -4,7 +4,7 @@ function checkExportData(currentList, exportData, apiName) {
       // check exportName and specifier
       if (specifier || exportName) {
         const defaultSpecifierName = specifier || exportName;
-        if (exportName === data.exportName || defaultSpecifierName === data.specifier) {
+        if ((exportName && exportName === data.exportName) || defaultSpecifierName === data.specifier) {
           throw new Error(`duplicate export data added by ${apiName},
             ${data.exportName ? `exportName: ${data.exportName}, ` : ''}specifier: ${data.specifier}
           `);
