+++ packages/plugin-helpers/src/index.ts
@@ -20,5 +20,4 @@ export default async function ({
   // .ice/index.ts:
   // export * from './helpers';
   applyMethod('addIceExport', { source: './helpers', exportName: 'helpers' });
-
 }
