+++ packages/plugin-request/src/types/index.ts
@@ -1,0 +1,6 @@
+import { AxiosRequestConfig } from 'axios';
+import { IInterceptors } from './base';
+
+export interface IRequest extends AxiosRequestConfig {
+  interceptors?: IInterceptors;
+}
