+++ packages/plugin-store/src/template/types.ts.ejs
@@ -1,4 +1,14 @@
-import * as types from '../appStore';
+import { IcestoreRootState, IcestoreDispatch } from '@ice/store';
+import { appModel } from '.';
 
-export type IRootDispatch = types.IRootDispatch;
-export type IRootState = types.IRootState;
+export type IRootDispatch = IcestoreDispatch<typeof appModel>;
+export type IRootState = IcestoreRootState<typeof appModel>;
+
+interface IInitialStates {
+  [key: string]: any;
+}
+
+export interface IStore {
+  initialStates?: IInitialStates;
+  getInitialStates?: (initialData) => IInitialStates;
+}
